+++ ui/src/themes/index.js
@@ -3,9 +3,11 @@ import DarkTheme from './dark'
 import ExtraDarkTheme from './extradark'
 import GreenTheme from './green'
 import SpotifyTheme from './spotify'
+import LigeraTheme from './ligera'
 
 export default {
   LightTheme,
+  LigeraTheme,
   DarkTheme,
   ExtraDarkTheme,
   GreenTheme,
